+++ database/migrations/2020_04_22_092513_create_module_instance_groupings_table.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateModuleInstanceGroupingsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     * 
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('module_instance_grouping', function (Blueprint $table) {
+            $table->bigIncrements('id');
+            $table->string('heading');
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('module_instance_grouping');    
+    }
+}