+++ docs/search/functions_7.js
@@ -82,5 +82,6 @@ var searchData=
   ['givepermissionto',['givePermissionTo',['../classBristolSU_1_1Support_1_1Testing_1_1HandlesAuthorization.html#aca35f7dbb376e3cb8566c6e171dbb03f',1,'BristolSU::Support::Testing::HandlesAuthorization']]],
   ['group',['group',['../classBristolSU_1_1Support_1_1Filters_1_1Contracts_1_1Filters_1_1GroupFilter.html#abf64b748e582669e96f1c792ed671c52',1,'BristolSU::Support::Filters::Contracts::Filters::GroupFilter']]],
   ['groupaudience',['groupAudience',['../classBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1Audience_1_1LogicAudience.html#a3ef19efa96a937744305f18f8415b839',1,'BristolSU::Support::Logic::Contracts::Audience::LogicAudience']]],
+  ['grouping',['grouping',['../classBristolSU_1_1Support_1_1ModuleInstance_1_1ModuleInstance.html#aa463e3a1fb6fba831e940bf32a381d41',1,'BristolSU::Support::ModuleInstance::ModuleInstance']]],
   ['groups',['groups',['../classBristolSU_1_1Support_1_1Logic_1_1Audience_1_1AudienceMember.html#a131ea8f66a516e79649c33b129a642b5',1,'BristolSU::Support::Logic::Audience::AudienceMember']]]
 ];